requirements.txt
matplotlib
numpy
Pillow
torch
torchvision
tqdm
yacs
scikit-learn